utils/requirements.txt
PyYAML>=6.0
nbformat>=5.7.0
jupyter-core>=5.0.0
